edit: Governor Namespace and floating point arithm [skip ci]
