+++ src/Model/PhasorDynamics/Governor/Tgov1/CMakeLists.txt
@@ -6,6 +6,6 @@
 
 gridkit_add_library(phasor_dynamics_governortgov1
   SOURCES
-    TurbineGov.cpp
+    Tgov1.cpp
   OUTPUT_NAME
     gridkit_phasor_dynamics_governortgov1)
